Updated layout and auth components
- updated css for table widget
- updated drawer size for ViewItemDetails.jsx component

Updated search for asset list

- update search button and placement for asset page
- update reports container ux.

Added changes to layout and auth components
